Switch to initialization option
