+++ metals/src/main/scala/scala/meta/internal/metals/MetalsLspService.scala
@@ -316,6 +316,7 @@ class MetalsLspService(
     tables,
     shellRunner,
     () => userConfig,
+    clientConfig,
   )
 
   private val bspConfigGenerator: BspConfigGenerator = new BspConfigGenerator(
